Moved function solar_angular_raduis() from map/maputils.py to coordinates/utils.py
